% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{GOMFPARENTS}
\alias{GOMFPARENTS}
\title{emulate the GO.db environments, but as functions: GOMF}
\usage{
GOMFPARENTS()
}
\value{
environment mapping from GOID to parents in MF ontology
}
\description{
emulate the GO.db environments, but as functions: GOMF
}
