% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{GOSYNONYM}
\alias{GOSYNONYM}
\title{emulate the GO.db environments, but as functions: GOSYNONYM}
\usage{
GOSYNONYM()
}
\value{
environment mapping from GOID to synonyms
}
\description{
emulate the GO.db environments, but as functions: GOSYNONYM
}
\examples{
head(unname(GOSYNONYM()[["GO:0009435"]]))
# this could be better
tmpcon = dbConnect(SQLite(), system.file("extdata", "go-basic.sqlite3", package="GO.db2"), flags=SQLITE_RO)
tbl(tmpcon, "go_synonym") |> dplyr::filter(go_id == "GO:0009435")
DBI::dbDisconnect(tmpcon)
}
