% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{superficial emulator of AnnotationDbi::select for Gene Ontology}
\usage{
select2(x, keys, columns = NULL, keytype, ...)
}
\arguments{
\item{x}{character, must be "GO.db2"}

\item{keys}{for filtering, if NULL, no filtering performed}

\item{columns}{for joining and selecting}

\item{keytype}{for table selection: one of "term", "goid"; all caps may be used}
}
\value{
data.frame
}
\description{
superficial emulator of AnnotationDbi::select for Gene Ontology
}
\note{
GO.db permitted columns and keytypes "DEFINITION", "GOID", "ONTOLOGY", "TERM"
}
\examples{
select2("GO.db2", keys="low-affinity zinc ion transmembrane transporter activity", keytype="term")
select2("GO.db2", keys="GO:0009435", keytype="goid") |> dplyr::select(-definition)
select2("GO.db2", keys="GO:0009435", keytype="goid", columns=c("GOID", "TERM", "ONTOLOGY"))
}
